pyproject.toml
[project]
name = "biomoqa-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.5.2",
    "arxiv>=2.1.3",
    "beautifulsoup4>=4.13.3",
    "biopython>=1.85",
    "bitsandbytes>=0.45.4",
    "datasets>=3.4.0",
    "deep-translator>=1.11.4",
    "dependencies>=7.7.1",
    "evaluate>=0.4.3",
    "faiss-cpu>=1.10.0",
    "hyperopt==0.2.5",
    "imblearn>=0.0",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.49",
    "lxml>=5.3.1",
    "matplotlib>=3.10.1",
    "mistralai>=1.6.0",
    "nlpaug>=1.1.11",
    "nltk>=3.9.1",
    "onnxruntime>=1.21.0",
    "openai>=1.68.2",
    "pandas>=2.2.3",
    "protobuf>=6.30.2",
    "pyalex>=0.18",
    "ray[default,tune]>=2.44.1",
    "regex>=2024.11.6",
    "runtime>=0.1.4",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.4.1",
    "sentencepiece>=0.2.0",
    "setuptools>=76.0.0",
    "tenacity>=9.0.0",
    "tensorboard>=2.19.0",
    "tensorboardx>=2.6.2.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
    "virtualenv>=20.29.3",
    "wordcloud>=1.9.4",
]
